requirements.txt
python-dotenv==0.21.0
requests==2.28.1
python-slugify==8.0.1
python-telegram-bot==11.1.0
redis==3.2.1
geopy==2.3.0